requirements.txt
torch>=1.9.0
torchvision~=0.11.3
numpy~=1.21.5
pandas~=1.3.5
matplotlib~=3.5.0
pillow~=8.4.0
tqdm~=4.62.3
ipython~=8.0.1
scikit-learn~=1.0.2